requirements.txt
selenium
pytest
allure-pytest
requests
faker
webdriver-manager
